requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python
cvzone==1.6.1
numpy
Pillow 